refactor: change directory structure.
