+++ src/pages/blog/[id].tsx
@@ -1,16 +1,16 @@
 import React from 'react';
 import { GetStaticPaths, GetStaticProps } from 'next';
-import { getArticleMetaData } from '../../src/libraries/articles';
-import { PageTemplate } from '../../src/templates/PageTemplate';
-import { ArticleMetaData } from '../../types/article';
-import { Optional } from '../../types/utility';
+import { getArticleMetaData } from '../../libraries/articles';
+import { PageTemplate } from '../../templates/PageTemplate';
+import { ArticleMetaData } from '../../../types/article';
+import { Optional } from '../../../types/utility';
 import styled from 'styled-components';
-import { color, fontSize } from '../../src/styles/variable';
+import { color, fontSize } from '../../styles/variable';
 import media from 'styled-media-query';
-import { TwitterShareButton } from '../../src/components/atoms/TwitterShareButton';
-import { GithubArticleButton } from '../../src/components/atoms/GithubArticleButton';
-import { MdTemplate } from '../../src/templates/MdTemplate';
-import { LinkComponent } from '../../src/components/atoms/LinkComponent';
+import { TwitterShareButton } from '../../components/atoms/TwitterShareButton';
+import { GithubArticleButton } from '../../components/atoms/GithubArticleButton';
+import { MdTemplate } from '../../templates/MdTemplate';
+import { LinkComponent } from '../../components/atoms/LinkComponent';
 
 export const config = { amp: true };
